public_html/requirements.txt
certifi==2019.9.11
Django==2.0.2
django-bootstrap-form==3.4
django-ckeditor==5.7.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-js-asset==1.2.2
django-responsive==0.3.0
django-user-agents==0.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
olefile==0.46
openpyxl==2.6.0
Pillow==6.0.0
pytz==2019.1
PyYAML==5.1.2
reportlab==3.5.31
sqlparse==0.3.0
ua-parser==0.8.0
user-agents==2.0
